+++ build.gradle
@@ -33,7 +33,6 @@ subprojects {
         annotationProcessor 'org.projectlombok:lombok'
 
         testImplementation 'org.springframework.boot:spring-boot-starter-test'
-        testImplementation 'org.springframework.security:spring-security-test'
         testRuntimeOnly 'com.h2database:h2'
     }
