--- winwin-be-domain/src/test/java/com/dpm/winwin/domain/configuration/TestConfig.java
@@ -1,20 +1,0 @@
-package com.dpm.winwin.domain.configuration;
-
-import com.dpm.winwin.domain.repository.post.impl.CustomPostRepositoryImpl;
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
-
-@TestConfiguration
-public class TestConfig {
-
-    @PersistenceContext
-    private EntityManager entityManager;
-
-    @Bean
-    public JPAQueryFactory jpaQueryFactory(){
-        return new JPAQueryFactory(entityManager);
-    }
-}
